moved PdvTokenizerHealthCheck type
